test create investment

+++ database/postgres/sqlc/models.go
@@ -203,17 +203,17 @@ type Investment struct {
 	AccountID     int64            `json:"account_id"`
 	Ticker        string           `json:"ticker"`
 	CompanyName   pgtype.Text      `json:"company_name"`
-	BuyVolume     int32            `json:"buy_volume"`
+	BuyVolume     int64            `json:"buy_volume"`
 	BuyValue      int64            `json:"buy_value"`
 	CapitalCost   int64            `json:"capital_cost"`
 	MarketPrice   int64            `json:"market_price"`
-	SellVolume    int32            `json:"sell_volume"`
+	SellVolume    int64            `json:"sell_volume"`
 	SellValue     int64            `json:"sell_value"`
-	CurrentVolume int32            `json:"current_volume"`
+	CurrentVolume int64            `json:"current_volume"`
 	Description   pgtype.Text      `json:"description"`
 	Status        InvestmentStatus `json:"status"`
-	Fee           int32            `json:"fee"`
-	Tax           int32            `json:"tax"`
+	Fee           int64            `json:"fee"`
+	Tax           int64            `json:"tax"`
 	UpdatedDate   pgtype.Timestamp `json:"updated_date"`
 }
 
@@ -223,13 +223,13 @@ type Transaction struct {
 	Ticker          string            `json:"ticker"`
 	TradingDate     pgtype.Timestamp  `json:"trading_date"`
 	Trade           TradeType         `json:"trade"`
-	Volume          int32             `json:"volume"`
+	Volume          int64             `json:"volume"`
 	OrderPrice      int64             `json:"order_price"`
-	MatchVolume     int32             `json:"match_volume"`
+	MatchVolume     int64             `json:"match_volume"`
 	MatchPrice      int64             `json:"match_price"`
 	MatchValue      int64             `json:"match_value"`
-	Fee             int32             `json:"fee"`
-	Tax             int32             `json:"tax"`
+	Fee             int64             `json:"fee"`
+	Tax             int64             `json:"tax"`
 	Cost            int64             `json:"cost"`
 	CostOfGoodsSold int64             `json:"cost_of_goods_sold"`
 	Return          int64             `json:"return"`
